move procedure to job
